+++ SDK/AppCenterDistribute/Microsoft.AppCenter.Distribute.Shared/DistributeDelegates.cs
@@ -9,6 +9,11 @@ namespace Microsoft.AppCenter.Distribute
     public delegate bool ReleaseAvailableCallback(ReleaseDetails releaseDetails);
 
     /// <summary>
+    /// App will close callback.
+    /// </summary>
+    public delegate void WillExitAppCallback();
+
+    /// <summary>
     /// No release available callback.
     /// </summary>
     public delegate void NoReleaseAvailableCallback();